Split Docker images into build and run images
Previously, there was a single Dockerfile defined that built Truss and
allowed users to run it. This yielded a tremendously large Docker image,
since much was downloaded and the entire Protobuf C++ implementation was
built inside it.

To reduce the size of the Docker image that most people will end up
using, one Dockerfile became two: one that builds an image with
everything needed to build Truss, and another that just runs Truss. When
the build image is run, the Truss source should be mounted into the
container at /go/src/github.com/TuneLab/go-truss, and the script that
gets run by the build image will build Truss into a build/ directory and
also copy Truss' runtime dependencies into build/.

Once this is complete, a runtime Truss image can be generated from the
contents of the build/ directory. The runtime image weighs in at about
25 MB, a fraction of the 1.6 GB combined image it replaces. The build
image is still around a gigabyte, but most users will not need to build
Truss.