Create build output directory outside of Docker container.
When running the truss-build image, output is placed in a build/
directory beneath the current directory. This is inherently created by
the "go build" command if it doesn't exist. Apparently, though, when
running on Linux, the Docker container's default root user can "bleed
out" into the host filesystem via a mounted volume. So, the "go build"
is being run as root, and it creates an output directory build/. When
the container finishes running, the resultant build/ directory will be
owned by root:root on the host system, not the user who ran the
Docker container.

We can avoid this by creating the build/ directory in the Makefile
before the Docker container is started. This kind of stinks, though,
because it means the user could end up in a bad situation if they run
the Docker container directly, outside the "control" of make. The best
we can do here is to check in the shell script that runs inside the
container whether the build directory already exists, and bail out if it
does not. Ugly, but more correct than allowing a user to end up in a bad
state with a root:root-owned build/ output directory.