Add gosu and wrapper script to allow non-root-owned output
Writing files to a bind-mounted volume on the host that is running
Docker that aren't owned by root requires running as a user that maps to
the UID and GID that the volume is mounted as. This is achieved using a
wrapper script that launches the actual build script, to create user and
group entries, run the build script as the new user, and change
ownership of everything inside of the bind-mounted directory after the
build script has completed.